+++ src/main/java/it/unimi/di/sweng/eventfinderbot/model/RequestHereToday.java
@@ -10,7 +10,7 @@ public class RequestHereToday extends Request {
     private Location location;
     private Date today;
 
-    public RequestHereToday(int chatId, Location location) {
+    public RequestHereToday(long chatId, Location location) {
         super(chatId);
         this.location = location;
         this.today = new Date();
